+++ wcomponents-theme/src/main/js/wc/dom/getLabelsForElement.js
@@ -69,6 +69,29 @@ define(["wc/dom/tag",
 		}
 
 		/**
+		 * Get element/s defined in aria-labelledby attribute
+		 * @param {Element} element The element with or without aria-labelledby attribute.
+		 * @returns {Array} An array of element/s. Element with 'id/s' listed in aria-labelledby attribute.
+		 * If either element with 'id' is not found, or aria-labelledby attribute is not found, then an empty array is returned.
+		 */
+		function getAriaLabelledElements(element) {
+			var ariaLabels = [],
+				labelIds = element.getAttribute("aria-labelledby");
+			if (labelIds) {
+				labelIds.split(/\s+/).forEach(function(labelId) {
+					var lblElement;
+					if (labelId) {
+						lblElement = document.getElementById(labelId);
+						if (lblElement) {
+							ariaLabels.push(lblElement);
+						}
+					}
+				});
+			}
+			return ariaLabels;
+		}
+
+		/**
 		 * Gets labelling element/s (label, legend or pseudo-label) for a control.
 		 *
 		 * @function module:wc/dom/getLabelsForElement
@@ -78,35 +101,35 @@ define(["wc/dom/tag",
 		 *  is returned.
 		 */
 		function getLabels(element, includeReadOnly) {
-			var result = [],
-				label,
-				tagName;
+			var ariaLabelledElements = [],
+				nativeLabeledElements = [],
+				label;
+			if (element) {
+				ariaLabelledElements = getAriaLabelledElements(element);
 
-			if (wrappedInput.isOneOfMe(element, includeReadOnly)) {
-				return getLabelsForWrapper(element, includeReadOnly);
-			}
+				if (wrappedInput.isOneOfMe(element, includeReadOnly)) {
+					nativeLabeledElements = getLabelsForWrapper(element, includeReadOnly);
+				} else {
+					FIELDSET = FIELDSET || new Widget(tag.FIELDSET);
+					if (FIELDSET.isOneOfMe(element)) {
+						LEGEND = LEGEND || new Widget("legend");
+						if ((label = LEGEND.findDescendant(element, true))) {
+							nativeLabeledElements = [label];
+						}
+					}
 
-			FIELDSET = FIELDSET || new Widget(tag.FIELDSET);
-			if (FIELDSET.isOneOfMe(element)) {
-				LEGEND = LEGEND || new Widget("legend");
-				if ((label = LEGEND.findDescendant(element, true))) {
-					result = [label];
+					nativeLabeledElements = doLabelQuery(element, nativeLabeledElements, includeReadOnly);
+
+					if (!(nativeLabeledElements && nativeLabeledElements.length)) {
+						// try getting an ancestor label element ONLY if element is input, textarea, select or progress.
+						if (LABELABLE.indexOf(element.tagName) > -1) {
+							nativeLabeledElements = getAncestorLabel(element);
+						}
+					}
 				}
+				// Append ariaLabel elements at the end of nativeLabel elements
+				return nativeLabeledElements.concat(ariaLabelledElements);
 			}
-
-			result = doLabelQuery(element, result, includeReadOnly);
-
-			if (result && result.length) {
-				return result;
-			}
-
-			// try getting an ancestor label element ONLY if element is input, textarea, select or progress.
-			tagName = element.tagName;
-			if (~LABELABLE.indexOf(tagName)) {
-				return getAncestorLabel(element);
-			}
-
-			return [];
 		}
 		/**
 		 * @module