+++ wcomponents-theme/src/test/intern/wc.dom.getLabelsForElement.test.js
@@ -44,7 +44,7 @@ define(["intern!object", "intern/chai!assert", "./resources/test.utils!"],
 			testGetLabelCountMoreThanOne: function() {
 				var element = document.getElementById("male"),
 					labels = controller(element),
-					expected = 2;
+					expected = 3;
 
 				assert.strictEqual(expected, labels.length);
 			},
@@ -66,6 +66,45 @@ define(["intern!object", "intern/chai!assert", "./resources/test.utils!"],
 				var result = controller(document.getElementById("unlabelled"));
 				assert.isTrue(Array.isArray(result));
 				assert.strictEqual(0, result.length);
+			},
+			testAriaLabel: function() {
+				var element = document.getElementById("aria-text"),
+					labels = controller(element),
+					expected = document.getElementById("aria-label");
+				assert.strictEqual(expected, labels[0]);
+			},
+			testAriaLabelUnavailable: function() {
+				var element = document.getElementById("aria-text2"),
+					labels = controller(element);
+				assert.strictEqual(0, labels.length);
+			},
+			testAriaMultipleLabel: function() {
+				var element = document.getElementById("aria-text3"),
+					labels = controller(element),
+					expected1 = document.getElementById("aria-label3"),
+					expected2 = document.getElementById("aria-label4");
+				assert.strictEqual(2, labels.length);
+				assert.strictEqual(expected1, labels[0]);
+				assert.strictEqual(expected2, labels[1]);
+			},
+			testAriaMultipleLabel2: function() {
+				var element = document.getElementById("aria-input"),
+					labels = controller(element),
+					expected1 = document.getElementById("billing"),
+					expected2 = document.getElementById("name");
+				assert.strictEqual(2, labels.length);
+				assert.strictEqual(expected1, labels[0]);
+				assert.strictEqual(expected2, labels[1]);
+			},
+			testDivElementAriaLabel: function() {
+				var element = document.getElementById("main"),
+					labels = controller(element),
+					expected = document.getElementById("foo");
+				assert.strictEqual(expected, labels[0]);
+			},
+			testNullElement: function() {
+				var labels = controller(null);
+				assert.isNotOk(labels);
 			}
 		});
 	});