+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### API Changes
 
 ### Enhancements
+* Client side: wc/dom/getLabelsForElement an enhancement to load elements defined by 'aria-labelledby' attribute #1401.
 
 ### Bug Fixes